Updated EDMv2Net with a magnitude preserving conditional branch
